+++ ocr/src/main/java/com/ing/software/ocr/OcrVars.java
@@ -5,8 +5,8 @@ package com.ing.software.ocr;
  */
 public class OcrVars {
 
-    static final boolean IS_DEBUG_ENABLED = true;
+    static final boolean IS_DEBUG_ENABLED = false;
     static final int LOG_LEVEL = 3; //A Higher level, means more things are logged
-    static final String[] AMOUNT_STRINGS = {"TOTALE", "IMPORTO", "AMOUNT"};
+    static final String[] AMOUNT_STRINGS = {"TOTALE", "IMPORTO"};
     static final int MAX_STRING_DISTANCE = 3;
 }
